feat: add render_indexed_content_section for indexes feature
Currently implementing render_section with indexes feature would require
writing an equivalent to `IndexedBasedReader`. As the logic is likely to
be the same all the time, I've added a `render_indexed_content_section`
function instead, to not expose the type and work on a
`ExactSizeIterator`. I'm not really sure if that's the best approach
though.

Furthermore there was an inconsistency, except `Vec<T>`, all other list
types like `&[T]` would simply ignore the `indexes` feature.